More calculation fixes.
+ corrected some edges not being properly added in the
community-community graph in the Louvain algorithm

+ added more debugging information being logged

NOT FIXING in this version, a math bug related to how hardware does
calculations that is aggravated by how the graph is implemented. Fixing
it requires reimplementation of the graph. This intermittent bug
sometimes causes community-community edges to remain instead of being
removed when the weight should be zero (but is not zero due to
aforementioned math bug).